FABRIK with constraints is very close to working.
The main issue - my best guess - is that if a joint reaches its limit, then its parent joint needs to be manipulated.

This could require some further analysis, perhaps worth exploring some A* algorithm so that it's not just maxing out one joint.

Also, the associated image is incorrect, but needed to envision what's going on. For now it will do.